+++ src/scraper/scraper.service.ts
@@ -2,17 +2,50 @@ import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 
 import Product from 'src/products/product.schema';
+import axios from 'axios';
 import { ProductService } from 'src/products/product.service';
-
-import { getAllMercadonaProducts } from './fetch/fetchMercadona';
+import { Category, CategoryResponse, Result, CategoryData, ProductData, PriceInstructions } from './interfaces/mercadona'
 
 @Injectable()
 export class ScraperService {
     constructor(private readonly productService: ProductService) { }
 
     async getMercadonaProducts(): Promise<void> {
-        const products: Product[] = await getAllMercadonaProducts();
-        products.forEach(product => {
+        const mercadonaCategoryList: Category[] = [];
+        const mercadonaProductList: Product[] = [];
+
+        // Get the full list of categories
+        const categoriesJSON: CategoryResponse = (await axios.get('https://tienda.mercadona.es/api/categories/')).data;        
+
+        categoriesJSON.results.forEach((upperCategory: Result) => {
+            upperCategory.categories?.forEach((lowerCategory: Result) => {
+                const category: Category = {id: lowerCategory.id, name: lowerCategory.name }
+                mercadonaCategoryList.push(category);
+            });
+        });
+
+        // Get All the info about all categories
+        const categoriesPromise: Array<Promise<CategoryData>> = mercadonaCategoryList.map(async (category: Category) => {
+            return axios.get(`https://tienda.mercadona.es/api/categories/${category.id}/?lang=es`);
+        });
+
+        const categories: CategoryData[] = await Promise.all(categoriesPromise);
+
+        // Fill the array with products
+        categories.forEach((data: CategoryData) => {
+            data.categories?.forEach((subCategory: CategoryData) => {
+                subCategory.products?.forEach((productData: ProductData) => {
+                    const product = new Product();
+                        product.name = productData.display_name;
+                        product.price = productData.price_instructions.unit_price;
+                        product.img = productData.thumbnail,
+                        product.description = "";
+                    mercadonaProductList.push(product);
+                });
+            });
+        });
+
+        mercadonaProductList.forEach(product => {
             this.productService.save(product)
         })
     }
